scripts refactor, leagues ids to variables
