+++ website/src/components/api-weather/Weather.vue
@@ -10,8 +10,8 @@
   import { defineComponent } from 'vue';
   import CurrentWeather from './CurrentWeather.vue';
   import Forecast from './Forecast.vue';
-  import { useWeather } from '@/scripts/WeatherService';
-  import { getWithExpiry } from "@/scripts/HandleItems.ts"
+  import { useWeather } from '@/scripts/Scripts.ts';
+  import { getWithExpiry } from "@/scripts/Scripts.ts";
   
   export default defineComponent({
     components: {
