--- website/src/scripts/CurrenciesModel.d.ts
@@ -1,12 +1,0 @@
-interface Rate {
-  currency: string;
-  code: string;
-  mid: number;
-}
-
-export interface ExchangeTable {
-  table: string;
-  no: string;
-  effectiveDate: string;
-  rates: Rate[];
-}
